Inject SCSSCacher for easier testing

+++ lib/private/Template/SCSSCacher.php
@@ -25,9 +25,12 @@ use Leafo\ScssPhp\Compiler;
 use Leafo\ScssPhp\Exception\ParserException;
 use Leafo\ScssPhp\Formatter\Crunched;
 use Leafo\ScssPhp\Formatter\Expanded;
+
+use OC\Files\AppData\Factory;
 use OCP\Files\IAppData;
 use OCP\Files\NotFoundException;
 use OCP\Files\NotPermittedException;
+use OCP\Files\SimpleFS\ISimpleFile;
 use OCP\Files\SimpleFS\ISimpleFolder;
 use OCP\ICache;
 use OCP\IConfig;
@@ -56,7 +59,7 @@ class SCSSCacher {
 
 	/**
 	 * @param ILogger $logger
-	 * @param IAppData $appData
+	 * @param Factory $appDataFactory
 	 * @param IURLGenerator $urlGenerator
 	 * @param IConfig $config
 	 * @param \OC_Defaults $defaults
@@ -64,14 +67,14 @@ class SCSSCacher {
 	 * @param ICache $depsCache
 	 */
 	public function __construct(ILogger $logger,
-								IAppData $appData,
+								Factory $appDataFactory,
 								IURLGenerator $urlGenerator,
 								IConfig $config,
 								\OC_Defaults $defaults,
 								$serverRoot,
 								ICache $depsCache) {
 		$this->logger = $logger;
-		$this->appData = $appData;
+		$this->appData = $appDataFactory->get('css');
 		$this->urlGenerator = $urlGenerator;
 		$this->config = $config;
 		$this->defaults = $defaults;
@@ -111,6 +114,16 @@ class SCSSCacher {
 	}
 
 	/**
+	 * @param $appName
+	 * @param $fileName
+	 * @return ISimpleFile
+	 */
+	public function getCachedCSS($appName, $fileName) {
+		$folder = $this->appData->getFolder($appName);
+		return $folder->getFile($fileName);
+	}
+
+	/**
 	 * Check if the file is cached or not
 	 * @param string $fileNameCSS
 	 * @param ISimpleFolder $folder
@@ -144,13 +157,11 @@ class SCSSCacher {
 
 	/**
 	 * Check if the variables file has changed
-	 * @param string $fileNameCSS
-	 * @param ISimpleFolder $folder
 	 * @return bool
 	 */
 	private function variablesChanged() {
 		$injectedVariables = $this->getInjectedVariables();
-		if($injectedVariables !== '' && $this->config->getAppValue('core', 'scss.variables') !== md5($injectedVariables)) {
+		if($this->config->getAppValue('core', 'scss.variables') !== md5($injectedVariables)) {
 			$this->resetCache();
 			$this->config->setAppValue('core', 'scss.variables', md5($injectedVariables));
 			return true;
@@ -230,9 +241,13 @@ class SCSSCacher {
 	 * We need to regenerate all files when variables change
 	 */
 	private function resetCache() {
-		foreach ($this->appData->getDirectoryListing() as $folder) {
+		$appDirectory = $this->appData->getDirectoryListing();
+		if(empty($appDirectory)){
+			return;
+		}
+		foreach ($appDirectory as $folder) {
 			foreach ($folder->getDirectoryListing() as $file) {
-				if(substr($file->getName(), -3) === "css" || substr($file->getName(), -4) === "deps") {
+				if (substr($file->getName(), -3) === "css" || substr($file->getName(), -4) === "deps") {
 					$file->delete();
 				}
 			}
